refactor(driverconfig): move to package
